requirements.txt
Flask>=2.3.3,<2.4.0
ngrok>=0.9.0,<0.10.0
PyAutoGUI>=0.9.54,<0.10.0
PyQRCode>=1.2.1,<1.3.0
python-dotenv>=1.0.0,<1.1.0
